+++ src/chess/game.rs
@@ -1,5 +1,5 @@
 use crate::chess::bitboard::bitboards;
-use crate::chess::movelist::MoveList;
+use crate::chess::moves::MoveList;
 use crate::chess::piece::Piece;
 use crate::chess::square::squares;
 use crate::chess::zobrist::ZobristHash;
@@ -135,7 +135,7 @@ impl Game {
         if self.halfmove_clock >= 100 {
             let mut movelist = MoveList::new();
             generate_legal_moves(self, &mut movelist);
-            return movelist.has_moves();
+            return !movelist.is_empty();
         }
 
         false