+++ src/engine/search/move_picker.rs
@@ -1,8 +1,7 @@
 use crate::chess::game::Game;
 use crate::chess::movegen;
 use crate::chess::movegen::MovegenCache;
-use crate::chess::movelist::MoveList;
-use crate::chess::moves::Move;
+use crate::chess::moves::{Move, MoveList};
 use crate::engine::search::move_ordering::score_move;
 use crate::engine::search::{move_ordering, PersistentState, SearchState};
 
@@ -135,7 +134,7 @@ impl MovePicker {
 
             if let Some(killer1) = state.killer_moves.get_0(plies) {
                 for i in self.first_quiet..self.moves.len() {
-                    if self.moves.get(i) == killer1 {
+                    if self.moves.get(i).map_or(false, |m| *m == killer1) {
                         self.moves.swap(self.first_quiet, i);
                         self.first_quiet += 1;
 
@@ -161,7 +160,7 @@ impl MovePicker {
 
             if let Some(killer2) = state.killer_moves.get_1(plies) {
                 for i in self.first_quiet..self.moves.len() {
-                    if self.moves.get(i) == killer2 {
+                    if self.moves.get(i).map_or(false, |m| *m == killer2) {
                         self.moves.swap(self.first_quiet, i);
                         self.first_quiet += 1;
 
@@ -226,7 +225,7 @@ impl MovePicker {
                 }
             }
 
-            let best_move = self.moves.get(best_move_idx);
+            let best_move = *self.moves.get(best_move_idx).unwrap();
 
             // Move our best move to the start of the moves we haven't tried
             self.moves.swap(self.idx, best_move_idx);
@@ -252,7 +251,11 @@ impl MovePicker {
         persistent_state: &PersistentState,
     ) {
         for i in start..end {
-            self.scores[i] = score_move(game, self.moves.get(i), &persistent_state.history_table);
+            self.scores[i] = score_move(
+                game,
+                *self.moves.get(i).unwrap(),
+                &persistent_state.history_table,
+            );
         }
     }
 }